LayoutTests/http/tests/security/contentSecurityPolicy/inline-style-blocked-expected.txt

CONSOLE MESSAGE: Refused to apply inline style because of Content-Security-Policy.

PASS
